+++ jarvis/api/triggers/stock_monitor.py
@@ -3,7 +3,6 @@
 import logging
 import math
 import os
-import sys
 from collections import defaultdict
 from datetime import datetime
 from typing import Dict, NoReturn, Union
@@ -13,13 +12,6 @@ import jinja2
 import matplotlib.dates
 import matplotlib.pyplot as plt
 from webull import webull
-
-sys.path.insert(0, os.getcwd())
-
-from jarvis.api.squire import stockmonitor_squire  # noqa
-from jarvis.modules.models import models  # noqa
-from jarvis.modules.templates import templates  # noqa
-from jarvis.modules.utils import util  # noqa
 
 
 def generate_graph(logger: logging.Logger, ticker: str, bars: int = 300) -> Union[str, NoReturn]:
@@ -236,9 +228,18 @@ class StockMonitor:
 
 
 if __name__ == '__main__':
+    # imports within __main__ to avoid potential/future path error and circular import
+    # override 'current_process().name' to avoid being set as 'MainProcess'
+    # importing at top level requires setting current_process().name at top level which will in turn override any import
+    from multiprocessing import current_process
+    current_process().name = "Crontab"
+    from jarvis.api.squire import stockmonitor_squire
     from jarvis.executors import crontab
     from jarvis.modules.logger import config
     from jarvis.modules.logger.custom_logger import logger as main_logger
+    from jarvis.modules.models import models
+    from jarvis.modules.templates import templates
+    from jarvis.modules.utils import util
 
     config.multiprocessing_logger(filename=crontab.LOG_FILE)
     for log_filter in main_logger.filters: