+++ jarvis/executors/others.py
@@ -7,9 +7,11 @@ import urllib.error
 import urllib.request
 from concurrent.futures import ThreadPoolExecutor
 from datetime import datetime
+from multiprocessing import current_process
 from threading import Thread
-from typing import List, NoReturn, Tuple
+from typing import Dict, List, NoReturn, Tuple, Union
 
+import boto3
 from googlehomepush import GoogleHome
 from googlehomepush.http_server import serve_file
 from joke.jokes import chucknorris, geek, icanhazdad, icndb
@@ -20,8 +22,8 @@ from pychromecast.error import ChromecastConnectionError
 from randfacts import get_fact
 
 from jarvis import version as module_version
-from jarvis.executors import (communicator, date_time, internet, robinhood,
-                              todo_list, weather, word_match)
+from jarvis.executors import (communicator, date_time, files, internet,
+                              robinhood, todo_list, weather, word_match)
 from jarvis.modules.audio import listener, speaker
 from jarvis.modules.conditions import keywords
 from jarvis.modules.database import database
@@ -33,6 +35,13 @@ from jarvis.modules.models import models
 from jarvis.modules.utils import shared, support, util
 
 db = database.Database(database=models.fileio.base_db)
+# set to accessible only via offline communicators
+# WATCH OUT: for changes in function name
+if current_process().name in ("fast_api", "telegram_api"):
+    SECRET_STORAGE = {'aws': [], 'local': []}
+    SESSION = boto3.Session()
+    SECRET_CLIENT = SESSION.client(service_name="secretsmanager")
+    SSM_CLIENT = SESSION.client(service_name="ssm")
 
 
 def repeat(phrase: str) -> NoReturn:
@@ -122,8 +131,8 @@ def music(phrase: str = None) -> NoReturn:
     Args:
         phrase: Takes the phrase spoken as an argument.
     """
-    get_all_files = (os.path.join(root, f) for root, _, files in os.walk(os.path.join(models.env.home, "Music")) for f
-                     in files)
+    get_all_files = (os.path.join(root, f) for root, _, file in os.walk(os.path.join(models.env.home, "Music")) for f
+                     in file)
     if music_files := [file for file in get_all_files if os.path.splitext(file)[1] == '.mp3']:
         chosen = random.choice(music_files)
         if phrase and 'speaker' in phrase:
@@ -428,3 +437,109 @@ def version() -> NoReturn:
         else:
             text += f", but the latest released version is {pkg_version}"
     speaker.speak(text=text)
+
+
+def get_aws_secrets(name: str = None) -> Union[Union[str, Dict[str, str]], List[str]]:
+    """Get secrets from AWS secretsmanager.
+
+    Args:
+        name: Get name of the particular secret.
+
+    Returns:
+        Union[Union[str, Dict[str]], List[str]]:
+        Returns the value of the secret or list of all secrets' names.
+    """
+    if name:
+        response = SECRET_CLIENT.get_secret_value(
+            SecretId=name
+        )
+        return response['SecretString']
+    paginator = SECRET_CLIENT.get_paginator('list_secrets')
+    page_results = paginator.paginate().build_full_result()
+    return [page['Name'] for page in page_results['SecretList']]
+
+
+def get_aws_params(name: str = None) -> Union[str, List[str]]:
+    """Get SSM parameters from AWS.
+
+    Args:
+        name: Get name of the particular parameter.
+
+    Returns:
+        Union[str, List[str]]:
+        Returns the value of the parameter or list of all parameter names.
+    """
+    if name:
+        response = SSM_CLIENT.get_parameter(Name=name, WithDecryption=True)
+        return response['Parameter']['Value']
+    paginator = SSM_CLIENT.get_paginator('describe_parameters')
+    page_results = paginator.paginate().build_full_result()
+    return [page['Name'] for page in page_results['Parameters']]
+
+
+def secrets(phrase: str) -> str:
+    """Handle getting secrets from AWS or local env vars.
+
+    Args:
+        phrase: Takes the phrase spoken as an argument.
+
+    Returns:
+        str:
+        Response to the user.
+    """
+    text = phrase.lower().split()
+
+    if 'list' in text:  # calling list will always create a new list in the dict regardless of what exists
+        if 'aws' in text:
+            SECRET_STORAGE['aws'] = []  # reset everytime list param is called
+            if 'ssm' in text:
+                try:
+                    SECRET_STORAGE['aws'].extend(get_aws_params())
+                except Exception as error:
+                    logger.error(error)
+            else:
+                try:
+                    SECRET_STORAGE['aws'].extend(get_aws_secrets())
+                except Exception as error:
+                    logger.error(error)
+            return ', '.join(SECRET_STORAGE['aws']) if SECRET_STORAGE['aws'] else "No parameters were found"
+        if 'local' in text:
+            SECRET_STORAGE['local'] = list(models.env.__dict__.keys())
+            return ', '.join(SECRET_STORAGE['local'])
+        return "Please specify which secrets you want to list: 'aws' or 'local''"
+
+    if 'get' in text:  # calling get will always return the latest information in the existing dict
+        if 'aws' in text:
+            if SECRET_STORAGE['aws']:
+                if aws_key := [key for key in phrase.split() if key in SECRET_STORAGE['aws']]:
+                    aws_key = aws_key[0]
+                else:
+                    return "No AWS params were found matching your request."
+            else:
+                return "Please use 'list secret' before using 'get secret'"
+            if 'ssm' in text:
+                try:
+                    key = util.keygen_uuid(length=16)
+                    files.put_secure_send(data={key: {aws_key: get_aws_params(name=aws_key)}})
+                    return key
+                except Exception as error:  # if secret is removed between 'list' and 'get'
+                    logger.error(error)
+            else:
+                try:
+                    key = util.keygen_uuid(length=16)
+                    files.put_secure_send(data={key: {aws_key: get_aws_secrets(name=aws_key)}})
+                    return key
+                except Exception as error:  # if secret is removed between 'list' and 'get'
+                    logger.error(error)
+            return f"Failed to retrieve {aws_key!r}"
+        if 'local' in text:
+            if not SECRET_STORAGE['local']:
+                SECRET_STORAGE['local'] = list(models.env.__dict__.keys())
+            if local_key := [key for key in phrase.split() if key in SECRET_STORAGE['local']]:
+                local_key = local_key[0]
+            else:
+                return "No local params were found matching your request."
+            key = util.keygen_uuid(length=16)
+            files.put_secure_send(data={key: {local_key: models.env.__dict__[local_key]}})
+            return key
+        return "Please specify which type of secret you want the value for: 'aws' or 'local'"