+++ jarvis/executors/port_handler.py
@@ -64,6 +64,6 @@ def kill_port_pid(port: int, protocol: str = 'tcp') -> Union[bool, None]:
     except (subprocess.SubprocessError, subprocess.CalledProcessError) as error:
         if isinstance(error, subprocess.CalledProcessError):
             result = error.output.decode(encoding='UTF-8').strip()
-            logger.error("[%d]: %s" % (error.returncode, result))
+            logger.error("[%d]: %s", error.returncode, result)
         else:
             logger.error(error)