+++ jarvis/executors/tv_controls.py
@@ -89,7 +89,7 @@ def tv_controller(phrase: str, tv_ip: str, identifier: str, nickname: str, clien
         elif 'open' in phrase_lower or 'launch' in phrase_lower:
             cleaned = ' '.join([w for w in phrase.split() if w not in ['launch', 'open', 'tv', 'on', 'my', 'the']])
             app_name = util.get_closest_match(text=cleaned, match_list=list(shared.tv[nickname].get_apps()))
-            logger.info("%s -> %s" % (phrase, app_name))
+            logger.info("%s -> %s", phrase, app_name)
             shared.tv[nickname].launch_app(app_name=app_name)
             speaker.speak(text=f"I've launched {app_name} on your {nickname} {models.env.title}!")
         elif "what's" in phrase_lower or 'currently' in phrase_lower:
@@ -98,7 +98,7 @@ def tv_controller(phrase: str, tv_ip: str, identifier: str, nickname: str, clien
             cleaned = ' '.join([word for word in phrase.split() if word not in ('set', 'the', 'source', 'on', 'my',
                                                                                 'of', 'to', 'tv')])
             source = util.get_closest_match(text=cleaned, match_list=list(shared.tv[nickname].get_sources()))
-            logger.info("%s -> %s" % (phrase, source))
+            logger.info("%s -> %s", phrase, source)
             shared.tv[nickname].set_source(val=source)
             speaker.speak(text=f"I've changed the source to {source}.")
         else: