+++ jarvis/executors/volume.py
@@ -27,7 +27,7 @@ def speaker_volume(level: int) -> NoReturn:
         level: Takes the volume level as an argument.
     """
     if models.audio_driver:
-        logger.info("Jarvis' volume has been set to %d" % level + "%")
+        logger.info("Jarvis' volume has been set to %d" % level + "%")  # % is mandatory because of string concatenation
         models.audio_driver.setProperty('volume', level / 100)
     else:
         logger.warning("Setting Jarvis' volume is not available for speech-synthesis")